chore(ts-migration): hits per page widget
